requirements.txt

pandas==1.3.5
xlrd==2.0.1
python-dotenv==0.21.1